requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
openai==1.51.0
requests==2.32.3
rapidfuzz==3.10.0
python-dotenv==1.0.1
python-multipart==0.0.13
pymupdf==1.24.9
